Optimized and modularized multiplayer services
